feat: move translation service initialization to see the logs
We don't really need to initialize it inside the server,
since server is only a way to use the service in the dev mode.

Logs of the server are written to a file due to it being started
from process which doesn't have access to console.
But we want to see TranslationService initialization logs
in the console, and the clearest way is to start it early,
rather than extracting validation into a separate function.